Do not fail compatibility check if error limited by maximumSeverity
Previously, if an issue was reported which was configured to have
`Severity.ERROR` then it would cause Metalava to fail with an
"Aborting: Found compatibility problems checking..." message. That is
even if `maximumSeverity` was lower than `Severity.ERROR`.

This change makes sure that if `maximumSeverity` limits a report to
less than `Severity.ERROR` then it does not cause Metalava to abort.

Bug: 311690008
Test: ./gradlew
Change-Id: I3d778d9f624eb489128122b27523073f5e7d181e