Add duplicate parameter to findCorrespondingItemIn(...)
When the `findCorrespondingItemIn(codebase, ...)` is called with
`superMethods` set to `true` then it can return a `MethodItem` from a
super class or interface. e.g. assume the `Codebase` currently being
operated on looks like this:

```
    public class Super {
        public void method() {...}
    }
    public final class Foo extends Super {
        public void method() {...}
    }
```

And the `codebase` being searched looks something like this (the `*`
suffix is just a way to differentiate classes in this `Codebase` from
the above `Codebase`):

```
    public class Super* {
        public void method() {...}
    }
    public final class Foo* extends Super {}
```

Then calling `findCorrespondingItemIn(..., superMethods=true)` on
`Foo.method()` will return `Super*.method()`. The former will be
considered `final` because it is in `Foo` which is `final`. However,
the latter will not be considered `final` as neither it or its
containing class `Super` is `final`. So, comparing them will lead
to incorrect results.

This change adds a `duplicate` parameter which if set to `true` when
calling it on `Foo.method()` will return `Foo*.method()` which will
also be considered `final` because the `Foo*` class is `final`. That
will ensure that comparisons work correctly.

Bug: 339166506
Test: ./gradlew
Change-Id: I22430ba13ab1f530e29512cc7854130ac20e4a55