Add a type visitor to handle multiple types
This pattern of simultaneously traversing multiple types assumed to have the same structure came up several times in creating checks for inner types.

This new visitor is essentially the same as the BaseTypeVisitor, but it also takes a list of other types in addition to the main type. When visiting an inner type of the main type, it also takes the corresponding inner type of each of the other types.

Bug: 332508325
Bug: 332507644
Test: N/A, usage tested in next change
Change-Id: Idf6bdb02d5adb7a35f44fa257b3b415a0dc78c5d