Refactor ModifierListWriter handling of final/open
The behavior of `final` and `open` differs between Java and Kotlin. In
Java everything is open by default (so does not support the `open`
keyword) whereas in Kotlin everything is `final` by default and has to
be explicitly declared to be `open`.

This change refactors the handling of the `final` and `open` modifiers
to make the differences between languages clearer.

Bug: 339166506
Test: ./gradlew
Change-Id: I3ceb7d00ba2c2a6fd10afd5e9cdea9b9f56916fe