Do not copy default modifier when copying to class
Previously, the `MethodItem.duplicate(ClassItem)` and
`ClassItem.inheritMethodFromNonApiAncestor(MethodItem)` methods would
copy the `default` modifier when copying an interface methods to a
class. That made no sense as the `default` modifier is only allowed on
interface methods. This change fixes that by adding a
`MethodItem.updateCopiedMethodState()` that clears the `default`
modifier in that case and which is called by all the implementations of
the above methods.

Bug: 339166506
Test: ./gradlew
Change-Id: I29c48f50d02f3a4a8523826dd35ccf91a482f7e4