Add test to show behavior when comparing method against super method
This change adds a test that shows the behavior when a concrete class
method is compared to a super default interface method in the
previously released API. They should be treated the same but the
`default` modifier results in the `UnflaggedApi` check thinking that
they have changed.

A follow-up change will use the `MethodItem.duplicate(ClassItem)`,
which has recently been updated to remove the `default` modifier from
an interface method that is copied into a concrete class.

Bug: 339166506
Test: ./gradlew
Change-Id: I3068a4794fcbbf0507b367d5c8a9fcaba1d545f3