Make TextMethodItem.duplicate(ClassItem) consistent with other models
Previously, `TextMethodItem.duplicate(ClassItem)` would copy the
`deprecated` property from the source method to the copied method. That
meant that if the source method was implicitly deprecated due to its
containing class being deprecated then the copied method would be
deprecated. However, if the source method we not deprecated then it
would not be implicitly deprecated in the target class even if that was
deprecated.

That is the opposite of the behavior of the other models which would
treat the copied method as being implicitly deprecated only if its
target class was deprecated.

This change makes the text model behavior consistent. It should not
result in changes in behavior downstream as signature files tend to
explicitly deprecate members of a deprecated class even if they were
not explicitly deprecated in the source.

Bug: 315206833
Test: ./gradlew
Change-Id: Idc0642de4f86b935b9f7786c343c93173c333404