Add test for how deprecated is handled for a removed class
Due to the use of the `ApiVisitor` when propagating the `deprecated`
status the contents of a removed class were not marked as `@Deprecated`
in the `removed-api.txt` file. That is because the `ApiVisitor` did not
visit the contents of a removed `Item`. This change adds a test to
ensure that follow-up changes do not break that behavior.

Longer term the inconsistencies in signature files will be cleaned up.

Bug: 315206833
Test: ./gradlew
Change-Id: Ifd11c06ff3fafd93c50101c344942e1984a32741