Make originallyDeprecated delegate to ModifierList.isDeprecated()
Previously, `originallyDeprecated` was a concrete read only property
that was initialized to `modifiers.isDeprecated()`. That meant that it
was impossible to change after it had been initialized. However, in
some cases, e.g. `ConvertJarsToSignatureFiles` it will be necessary to
change that.

So, to fix that this change switches `originallyDeprecated` to be an
abstract property that simply returns the value of
`modifiers.isDeprecated()` instead of taking a snapshot so that changes
to the `ModifierListWriter.isDeprecated()` state will affect the value
of `originallyDeprecated` and thus the value of
`effectivelyDeprecated`.

Bug: 315206833
Test: ./gradlew
Change-Id: I7751f73fac83474350600e87b3ffaed6b49983ab