Remove unnecessary deprecation attributes in api-versions.xml
Previously, members of a deprecated class would have a "deprecated"
attribute that reported that they were deprecated in the same API
version as the class (if they were not deprecated prior to that). That
is unnecessary as the code that consumes the `api-versions.xml` file
will use the class's deprecated API version for members if the members
do not specify their own version.

This change will only add the "deprecated" attribute if it is
different to the parent element's reducing the size of the generated
`api-versions.xml` file.

Bug: 315206833
Test: ./gradlew
Change-Id: I7affaf7a32625e51b5635225e30446dddcc3d34f