Move super class filtering from SignatureWriter
Previously, super class filtering was performed by `SignatureWriter`.
This change moves that into the `FilteringClassItem` in
`FilteringApiVisitor`. It also adds a `preFiltered` property as that is
needed to ensure identical behavior.

The `writeExtendsOrImplementsType(ClassTypeItem)` method passes
`filterReference` to `TypeItem.toTypeString(...)`'s `filter` property.
That means that it will exclude any type annotation from the output
that is resolvable but for whose `ClassItem` `filterReference` returns
`false`. In order to replicate that behavior it was necessary to add a
`TypeTransformer` that will do the same for the super class type.

The `writeExtendsOrImplementsType(ClassTypeItem)` method is unchanged
by this change because it is also used for writing out interface types.

Bug: 348701068
Test: ./gradlew
Change-Id: I2b32ee3a59c6c0d1e7514cf98faf2114a0bce121