Remove special handling of getSystemService(...)
Previously, Metalava would remove the `@Nullable` annotation from any
`getSystemService(...)` method to prevent developers from having to
handle the method returning `null` for services that are guaranteed to
exist. However, that hack made the resulting API stubs less correct and
could lead to runtime NPEs. This change removes the hack.

That was the only use of `Issues.EXPECTED_PLATFORM_TYPE` so this
removes that as well.

Bug: 347269120
Test: ./gradlew
Change-Id: Idc28559ccb22ed3d842475c11b504a31e926eba7