Make nullness migration of system API test more realistic
Previously, it would provide the previous current and system APIs in
one signature file. That is not realistic as they would usually be
provided as separate files. This changes the test to supply them as
separate files which causes a regression in the behavior.

A follow-up change will fix the test.

Bug: 347885819
Test: ./gradlew
Change-Id: I35dede486b913ba440f987029b356a2e62f9bd7a