Normalize indenting/formatting of NullnessMigrationTest
The previous indenting made it difficult to separate the content of the
different files. The formatting also did not adding `,` to the last
argument of methods that can take multiple arguments which makes future
changes that add arguments cause more churn as they have to add a comma
on the previous line. Finally, there was an empty `expectedIssues`
parameter that was not needed.

This change fixes all the above.

Bug: 347885819
Test: ./gradlew
Change-Id: I54bdf93cbb5322287dbf7c8201247563f646a882