Update the PackageItem.emit property correctly
Previously, a `PackageItem` that was defined in a `SignatureFile` for
which `forCurrentApiSurface` was `false` would have its `emit` property
set to `false` and that would not be changed even if it was also
defined in another signature file that set `forCurrentApiSurface` to
`true`. That would prevent the classes in that package from being
visited by an `ApiVisitor`, including `CodebaseComparator`.

This change fixes the issue and fixes the test that was previously
testing the behavior.

This change revealed a separate issue with the handling of overlapping
methods, i.e. methods that are defined in multiple signature files.
That will be fixed in a follow-up.

Bug: 347922041
Test: ./gradlew
Change-Id: Ice0e23d7a8c4ed250e3ed956b427eaa7d5664d9e