Add support for specifying annotations important for API compatibility
This adds support for a new option "--api-compat-annotation". This
argument will accept a fully qualified annotation name and supports
multiple instances. These annotations will be treated as important to
API compatibility and will fail compatibility checks if they are
added/removed.

A good use-case would be the @Composable annotation, where
adding/removing it from a target function is a breaking change, due to
the modifications done by the Compose Compiler.

Bug: None
Test: ./gradlew

Change-Id: If85229ae536e2ae6cffe0ffee73c5d300dc2b841