Pull call to normalizeInputName(String?) out of constructor
Previously, `AnnotationManager.normalizeInputName(String?): String?`
was called from the `DefaultAnnotationItem` constructor to map the
`originalName` to the `qualifiedName`. As it could return null it
forced the `qualifiedName` to be nullable.

This change pushed the call to that method out of the constructor
and into factory methods. Follow-up changes can then make those factory
methods return `null` when the `originalName` or `qualifiedName` is
`null` which will in turn allow `AnnotationItem`'s `originalName` and
`qualifiedName` to be changed from `String` to `String?`.

Bug: 347944677
Test: ./gradlew
Change-Id: Ib768c8721db34890154f2b8fa3cf8a9deb587598