Add test for AnnotationItem.fileLocation
Adds `AnnotationItem.fileLocation` which, by default returns
`FileLocation.UNKNOWN` and some tests to show the behavior of that for
Java and Kotlin. Signature file locations are not so important as they
are always on the same line as the annotated item.

At the moment the locations of the annotations do not appear in the
test because they are unknown but follow-up changes will fix that for
the Psi and Turbine models.

Bug: 346530869
Test: ./gradlew
Change-Id: I93c79fe62a3f8908d328c494d41626064b094c89