Use whole API surface when migrating nullability annotations
The preceding change added a test that showed when the previously
released codebase that is used for migrating nullness is provided by
two separate files (e.g. public and system) then fields that are
defined in the first file do not have their nullability migrated
properly. That is because when loading the files the items defined in
the first file have `emit` set to `false` because the corresponding
`SignatureFile` has `forCurrentApiSurface` set to `false`.

Methods are not affected because there is a "feature" in
`ComparisonVisitor` where it will "inherit" a method from the same
class and then ignore its `emit` property. That will be dealt with
separately.

This change fixes that by adding an `onlyUseLastForCurrentApiSurface`
parameter that can be used to ensure that all files loaded for the
previously released API have `forCurrentApiSurface` and hence `emit`
set to `true`.

Bug: 347885819
Test: ./gradlew
Change-Id: I54cf362945eb4bd2a6d6a2d12c985af90ef306fa