Remove --check-compatibility:base
The `--check-compatibility:base` option was a bit of a hack that was
added because `--check-compatibility:api:released` did not support
being specified multiple times. The only uses of the former have been
replaced with the latter and so it can now be removed.

Removing it allows a lot of other code to be cleaned as it was the only
reason why `ComparisonVisitor` needed to support handling comparing two
lists of `Codebase`s. Removing that allowed removal of `MergedCodebase`
as well `removeDuplicates(ItemTree)`.

A number of tests that either directly tested the behavior of the base
`Codebase`, or relied on it have been deleted after first checking
whether there was any point to migrating them to use multiple copies of
`--check-compatibility:api:released`.

Bug: 320245343
Test: ./gradlew
Change-Id: I151bdb150b3d8e3fddb7049796cb197f575eb5b7