Improve NullnessMigrationTest for multiple previously released APIs
Previously, the test only checked methods on an interface. This change
adds tests for constructor and field in public and system concrete
classes. That reveals an issue with the handling of fields in the
public but not system classes. The field in the public class does not
have its `@Nullability` annotation replaced with `@RecentlyNullable`
but the field in the system class does.

A follow-up change will fix that.

Bug: 347885819
Test: ./gradlew
Change-Id: If79b5f782079e218972d99850f7d8131e2008b31