Add ComparisonVisitor test to show behavior when method has emit=false
The test shows that when a method exists but has `emit=false` and is
compared against the same method with `emit=true` that
`ComparisonVisitor` does not ignore the method but instead treats it
as if it was changed. That is not the expected behavior and is not the
behavior for fields. A follow-up change will correct that.

Bug: 347885819
Test: ./gradlew
Change-Id: Ie22d36030667025c3401c9f1c64d963dda0602e6