Limit gathering of API jar files to --current-version
Previously, `findAndroidJars(...)` would search for all jars from the
`--first-version` until the last jar file that could be found. That
caused the `ExtractSystemApiLevelsTest` to start failing when the
`prebuilts/sdk/35` was dropped.

This change limits the search up to and including `--current-version`
which fixes the test which already sets `--current-version` to 32.

Bug: 341941365
Test: ./gradlew
      # Merge into AOSP and make sure it does not change the contents
      # of the following files:
      # * out/soong/lint/api_versions_public.xml
      # * out/soong/lint/api_versions_system.xml
      # * out/soong/lint/api_versions_module_lib.xml
      # * out/soong/lint/api_versions_system_server.xml
Change-Id: I8d47aa1d54ccdd0fd761ab42171703ff5d3bff2e