Separate duplicate(TypeModifiers) and substitute(TypeModifiers)
Previously, `TypeItem.duplicate(TypeModifiers)` was provided to allow
the `TypeItem.modifiers` to be changed without having to handle each
subinterface specially. However, that could lead to it creating
unnecssary instances. It was also implemented by the concrete classes
for those `TypeItem` subinterfaces that did not have their own
`duplicate(...)` method, i.e. `PrimitiveTypeItem` and
`VariableTypeItem`.

This change separates those two uses. The `duplicate(TypeModifiers)` is
pushed down into `PrimitiveTypeItem` and `VariableTypeItem` and the
`TypeItem.duplicate(TypeModifiers)` (and overrides to specialize the
return type) are renamed to `substitute` and either call the existing
`substitute(...)` method or in the case of the `PrimitiveTypeItem` and
`VariableTypeItem` types, uses their `duplicate(TypeModifiers)` method
directly.

Bug: 325454211
Test: ./gradlew
Change-Id: Ieb8598ec333464c4a2f19235f08cbb86565f8e4e