Deprecate TypeItem's duplicate(...) methods as implementation detail
The various `duplicate(...)` methods are only needed for use by the
corresponding `substitute(...)` methods to delegate the work of
creating a new instance to the implementation class. It avoids having
to implement the checking logic in `substitute(...)` in multiple
implementation classes.

This change makes that clear by deprecating the methods in favor of the
corresponding `substitute(...)` method.

Bug: 325454211
Test: ./gradlew
Change-Id: Ib379e2ea910f8478fc191c8f1ce264e44e61757f