Add InterfaceTypeListOrderTest to improve test coverage
While some tests will break if the `SignatureWriter` does not work
correctly when `FileFormat.sortWholeExtendsList=false` there is at
least one case where the tests do not fail. That is when the first
type in the interface's `extends` list is a private interface. This
change adds a test to cover that.

Bug: 348701068
Test: ./gradlew
Change-Id: Ic6025290473add98e7a396f898b55f6db5dc2bc4